requirements.txt
absl-py==0.11.0
astunparse==1.6.3
backcall==0.2.0
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
decorator==4.4.2
Flask==1.1.2
Flask-Compress==1.8.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.2.0
parso==0.8.1
pdf2image==1.14.0
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plotly==4.14.3
prompt-toolkit==3.0.10
protobuf==3.15.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF4==1.27.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.5
pyzmq==21.0.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.1
torchvision==0.8.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
